implemented implemented ExceptionHandler magazine
