+++ src/main/java/com/teamcubation/librarymanagement/adapter/in/web/exceptionhandling/ControllerExceptionHandling.java
@@ -1,5 +1,7 @@
 package com.teamcubation.librarymanagement.adapter.in.web.exceptionhandling;
 
+import com.teamcubation.librarymanagement.domain.exceptions.magazine.MagazineAttributeMissingException;
+import com.teamcubation.librarymanagement.domain.exceptions.magazine.MagazineNotAvailableException;
 import com.teamcubation.librarymanagement.domain.exceptions.room.RoomAlreadyExistsException;
 import com.teamcubation.librarymanagement.domain.exceptions.room.RoomAlreadyReservedException;
 import com.teamcubation.librarymanagement.domain.exceptions.room.RoomNotFoundException;
@@ -25,5 +27,14 @@ public class ControllerExceptionHandling {
     @ExceptionHandler(RoomNotFoundException.class)
     public ResponseEntity<String> handleRoomNotFoundException() {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("No se encontro la sala");
+    }
+    @ExceptionHandler({MagazineAttributeMissingException.class})
+    public ResponseEntity<String> handleExceptionMagazineAttributeMissingException() {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Your magazine attributes are void");
+    }
+
+    @ExceptionHandler({MagazineNotAvailableException.class})
+    public ResponseEntity<String> handleExceptionMagazineNotAvailableException() {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Your magazine is not available");
     }
 }
